+++ beachball.config.js
@@ -1,3 +1,3 @@
 require('./scripts/ts-node-register');
 
-module.exports = require('./scripts/beachball.config').config;
+module.exports = require('./scripts/beachball/index').config;
