+++ scripts/beachball/index.ts
@@ -1,0 +1,18 @@
+import { BeachballConfig } from 'beachball';
+import { renderHeader, renderEntry } from './customRenderers';
+
+export const config: BeachballConfig = {
+  groups: [
+    {
+      name: 'Fluent UI React',
+      include: ['packages/office-ui-fabric-react', 'packages/react'],
+      disallowedChangeTypes: ['major'],
+    },
+  ],
+  changelog: {
+    customRenderers: {
+      renderHeader,
+      renderEntry,
+    },
+  },
+};
