+++ scripts/package.json
@@ -64,7 +64,7 @@
     "babel-jest": "^24.5.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-lodash": "^3.3.4",
-    "beachball": "^1.20.4",
+    "beachball": "^1.28.0",
     "chalk": "^2.1.0",
     "circular-dependency-plugin": "^5.0.2",
     "clean-webpack-plugin": "^0.1.19",
